+++ library/src/plan.cpp
@@ -31,6 +31,7 @@
 
 #include <algorithm>
 #include <assert.h>
+#include <functional>
 #include <map>
 #include <numeric>
 #include <set>
@@ -711,8 +712,20 @@ ROCFFT_EXPORT rocfft_status rocfft_repo_get_total_plan_count(size_t* count)
 // Factorisation helpers
 //
 
-inline size_t get_explicity_supported_factor(rocfft_precision precision, size_t length0)
+// Return true if the order of factors in a decomposition should be
+// reversed.  This improves performance for some lengths.
+inline bool reverse_factors(size_t length)
 {
+    std::set<size_t> reverse_factors_lengths = {32256, 43008};
+    return reverse_factors_lengths.count(length) == 1;
+}
+
+// Search function pool for length where is_supported_factor(length) returns true.
+inline size_t search_pool(rocfft_precision            precision,
+                          size_t                      length,
+                          std::function<bool(size_t)> is_supported_factor)
+{
+    // query supported lengths from function pool, largest to smallest
     auto supported  = function_pool::get_lengths(precision, CS_KERNEL_STOCKHAM);
     auto comparison = std::greater<size_t>();
     std::sort(supported.begin(), supported.end(), comparison);
@@ -720,43 +733,44 @@ inline size_t get_explicity_supported_factor(rocfft_precision precision, size_t 
     if(supported.empty())
         return 0;
 
-    size_t factor = 0;
+    // start search slightly smaller than sqrt(length)
+    auto v     = (size_t)sqrt(length);
+    auto lower = std::lower_bound(supported.cbegin(), supported.cend(), v, comparison);
+    if(*lower < sqrt(length))
+        lower--;
 
-    if(length0 > (Large1DThreshold(precision) * Large1DThreshold(precision)))
-    {
-        auto supported_factor = [length0](size_t factor) -> bool {
-            bool is_factor = length0 % factor == 0;
-            return is_factor;
-        };
+    auto upper = supported.cend();
 
-        auto v     = Large1DThreshold(precision);
-        auto lower = std::lower_bound(supported.cbegin(), supported.cend(), v, comparison);
-        auto itr   = std::find_if(lower, supported.cend(), supported_factor);
-        if(itr != supported.cend())
-            factor = *itr;
-    }
-    else
-    {
-        // break into as squarish matrix as possible
-        auto supported_factor = [length0, precision = precision](size_t factor) -> bool {
-            bool is_factor    = length0 % factor == 0;
-            bool have_kernels = function_pool::has_function(fpkey(length0 / factor, precision));
-            return is_factor && have_kernels;
-        };
+    // search!
+    auto itr = std::find_if(lower, upper, is_supported_factor);
+    if(itr != supported.cend())
+        return *itr;
 
-        auto v     = (size_t)sqrt(length0);
-        auto lower = std::lower_bound(supported.cbegin(), supported.cend(), v, comparison);
-        if(*lower < sqrt(length0))
-            lower--;
+    return 0;
+}
 
-        // note: this squarish factorization isn't always the fastest.
-        // for length 18816: if you start at supported.cbegin() the resulting plan is faster
-        auto itr = std::find_if(lower, supported.cend(), supported_factor);
-        if(itr != supported.cend())
-            factor = *itr;
-    }
-
+// Return largest factor that has BOTH functions in the pool.
+inline size_t get_explicitly_supported_factor(rocfft_precision precision, size_t length)
+{
+    auto supported_factor = [length, precision = precision](size_t factor) -> bool {
+        bool is_factor        = length % factor == 0;
+        bool has_other_kernel = function_pool::has_function(fpkey(length / factor, precision));
+        return is_factor && has_other_kernel;
+    };
+    auto factor = search_pool(precision, length, supported_factor);
+    if(factor > 0 && reverse_factors(length))
+        return length / factor;
     return factor;
+}
+
+// Return largest factor that has a function in the pool.
+inline size_t get_largest_supported_factor(rocfft_precision precision, size_t length)
+{
+    auto supported_factor = [length](size_t factor) -> bool {
+        bool is_factor = length % factor == 0;
+        return is_factor;
+    };
+    return search_pool(precision, length, supported_factor);
 }
 
 inline bool SupportedLength(rocfft_precision precision, size_t len)
@@ -765,8 +779,7 @@ inline bool SupportedLength(rocfft_precision precision, size_t len)
     if(function_pool::has_function(fpkey(len, precision)))
         return true;
 
-    // can we factor with 2, 3, or 5?  note: all combinations of these
-    // are explicitly generated at build time
+    // can we factor with using only base radix?
     size_t p = len;
     while(!(p % 2))
         p /= 2;
@@ -774,6 +787,12 @@ inline bool SupportedLength(rocfft_precision precision, size_t len)
         p /= 3;
     while(!(p % 5))
         p /= 5;
+    while(!(p % 7))
+        p /= 7;
+    while(!(p % 11))
+        p /= 11;
+    while(!(p % 13))
+        p /= 13;
 
     if(p == 1)
         return true;
@@ -783,7 +802,7 @@ inline bool SupportedLength(rocfft_precision precision, size_t len)
         return true;
 
     // finally, can we factor this length with combinations of existing kernels?
-    if(get_explicity_supported_factor(precision, len) > 0)
+    if(get_explicitly_supported_factor(precision, len) > 0)
         return true;
 
     return false;
@@ -1629,15 +1648,6 @@ void TreeNode::build_real_even_3D()
     }
 }
 
-size_t TreeNode::div1DNoPo2(const size_t length0)
-{
-    auto factor = get_explicity_supported_factor(precision, length0);
-    // return 0 means we can't find any supported kernels,
-    // happens when debugging, we don't want this assert crash the program
-    //assert(factor != 0);
-    return (factor > 0) ? length0 / factor : 0;
-}
-
 // Compute the large twd decomposition base
 // 2-Steps:
 //  e.g., ( CeilPo2(10000)+ 1 ) / 2 , returns 7 : (2^7)*(2^7) = 16384 >= 10000
@@ -1663,9 +1673,7 @@ void TreeNode::build_1D()
         return;
     }
 
-    // single: limit up to 4096, double: up to 2048
-    if(length[0] <= Large1DThreshold(precision)
-       && function_pool::has_function(fpkey(length[0], precision)))
+    if(function_pool::has_function(fpkey(length[0], precision)))
     {
         scheme = CS_KERNEL_STOCKHAM;
         return;
@@ -1677,10 +1685,13 @@ void TreeNode::build_1D()
     if(IsPo2(length[0])) // multiple kernels involving transpose
     {
         // TODO: wrap the below into a function and check with LDS size
-        if(length[0] <= 262144 / PrecisionWidth(precision))
+        auto block_threshold = 262144;
+        if(precision == rocfft_precision_double)
+            block_threshold /= 2;
+        if(length[0] <= block_threshold)
         {
             // Enable block compute under these conditions
-            if(1 == PrecisionWidth(precision))
+            if(precision == rocfft_precision_single)
             {
                 if(map1DLengthSingle.find(length[0]) != map1DLengthSingle.end())
                 {
@@ -1704,13 +1715,13 @@ void TreeNode::build_1D()
             }
             // up to 256cc + 256rc for the 1arge-1D, use CRT for even larger
             scheme = (length[0] <= 65536) ? CS_L1D_CC : CS_L1D_CRT;
-            // scheme = (length[0] <= 65536 / PrecisionWidth(precision)) ? CS_L1D_CC : CS_L1D_CRT;
         }
         else
         {
-            if(length[0] > (Large1DThreshold(precision) * Large1DThreshold(precision)))
+            auto largest = function_pool::get_largest_length(precision);
+            if(length[0] > largest * largest)
             {
-                divLength1 = length[0] / Large1DThreshold(precision);
+                divLength1 = length[0] / largest;
             }
             else
             {
@@ -1756,9 +1767,19 @@ void TreeNode::build_1D()
 
         if(failed)
         {
-            divLength1 = div1DNoPo2(length[0]);
             scheme     = CS_L1D_TRTRT;
-            failed     = (divLength1 == 0);
+            divLength1 = get_explicitly_supported_factor(precision, length[0]);
+            if(divLength1 == 0)
+            {
+                // We need to recurse.  Note, for CS_L1D_TRTRT,
+                // divLength0 has to be explictly supported
+                auto divLength0 = get_largest_supported_factor(precision, length[0]);
+                if(divLength0 == 0)
+                    divLength1 = 0;
+                else
+                    divLength1 = length[0] / divLength0;
+            }
+            failed = divLength1 == 0;
         }
     }
 
@@ -1879,12 +1900,6 @@ void TreeNode::build_1DBluestein()
 
 void TreeNode::build_1DCS_L1D_TRTRT(const size_t divLength0, const size_t divLength1)
 {
-    if(!function_pool::has_function(fpkey(divLength0, precision)))
-    {
-        PrintFailInfo(precision, length[0], scheme, divLength0, CS_KERNEL_STOCKHAM);
-        assert(false);
-    }
-
     // first transpose
     auto trans1Plan = TreeNode::CreateNode(this);
 
@@ -1954,7 +1969,11 @@ void TreeNode::build_1DCS_L1D_TRTRT(const size_t divLength0, const size_t divLen
     }
 
     // algorithm is set up in a way that row2 does not recurse
-    assert(divLength0 <= Large1DThreshold(this->precision));
+    if(!function_pool::has_function(fpkey(divLength0, precision)))
+    {
+        PrintFailInfo(precision, divLength0, scheme);
+        assert(false);
+    }
 
     childNodes.emplace_back(std::move(row2Plan));