+++ library/src/powX.cpp
@@ -403,8 +403,7 @@ bool PlanPowX(ExecPlan& execPlan)
             assert(false);
         }
 
-        gp.lds_bytes = (lds + lds_padding * bwd) * PrecisionWidth(execPlan.execSeq[0]->precision)
-                       * sizeof(float2);
+        gp.lds_bytes = (lds + lds_padding * bwd) * sizeof_precision(execPlan.execSeq[0]->precision);
         execPlan.execSeq[i]->lds_padding = lds_padding;
         execPlan.devFnCall.push_back(ptr);
         execPlan.gridParam.push_back(gp);